program displays all words similar to user's input
accepts a file name
from command line containing a sorted list of words. These sorted list
of
words are then loaded and their substrings are created. Then the program

asks user to enter a word to find a similar list of words in the
dictionary.
Program also creates substrings of that word and then compares
substrings
of user's words and words in dictionary. For each word in the dictionary
if
any of substrings matches with that of user's word, then that dictionary

word is outputed. The program asks user to input a word until user
declines to do so.